requirements.txt
geopandas==0.9.0
netCDF4==1.4.2
numpy==1.16.4
networkx==2.3.0
scikit-image==0.17.2
matplotlib==3.1.0
shapely==1.6.4